% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_name_to_statistics.R
\name{add_name_to_statistics}
\alias{add_name_to_statistics}
\alias{add_name_to_statistics.list}
\alias{add_name_to_statistics.data.frame}
\title{add_name_to_statistics}
\usage{
add_name_to_statistics(x, name, ...)

\method{add_name_to_statistics}{list}(x, name, ...)

\method{add_name_to_statistics}{data.frame}(
  x,
  name,
  colname_for_variable = atable_options("colname_for_variable"),
  ...
)
}
\arguments{
\item{x}{an object}

\item{name}{a value}

\item{...}{passed to methods}

\item{colname_for_variable}{a character length 1. Default is defined in atable_options}
}
\value{
x now with new field colname_for_variable
}
\description{
add a column to a data.frame x with value name as character.
Helper Function. Not intended to be called by the user.
}
\details{
checks if the new field already exists
}
\section{Methods (by class)}{
\itemize{
\item \code{add_name_to_statistics(list)}: apply add_name_to_statistics to all field of the list

\item \code{add_name_to_statistics(data.frame)}: add field colname_for_variable to the data.frame. chekc for a name clash as this field as there are many user-defined fields

}}
