% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_name_to_tests.R
\name{add_name_to_tests}
\alias{add_name_to_tests}
\alias{add_name_to_tests.list}
\alias{add_name_to_tests.data.frame}
\title{add_name_to_tests}
\usage{
add_name_to_tests(x, name, ...)

\method{add_name_to_tests}{list}(x, name, ...)

\method{add_name_to_tests}{data.frame}(
  x,
  name,
  colname_for_variable = atable_options("colname_for_variable"),
  ...
)
}
\arguments{
\item{x}{an object}

\item{name}{a value}

\item{...}{passed to methods}

\item{colname_for_variable}{a character length 1. Default is defined in atable_options}
}
\value{
x now with new field colname_for_variable
}
\description{
Helper-function to add a field to a list or data.frame
}
\details{
Not intended to be called by the user.

checks if the new field already exists
}
\section{Methods (by class)}{
\itemize{
\item \code{add_name_to_tests(list)}: apply add_name_to_statistics to all field of the list

\item \code{add_name_to_tests(data.frame)}: add field colname_for_variable to the data.frame. chekc for a name clash as this field as there are many user-defined fields

}}
