% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_observation_column.R
\name{add_observation_column}
\alias{add_observation_column}
\title{Adds a column to a data.frame}
\usage{
add_observation_column(DD)
}
\arguments{
\item{DD}{A data.frame.}
}
\value{
As DD now with one more column.
}
\description{
The new column has name \code{atable_options('colname_for_observations')} and class \code{'count_me'}.
}
\details{
Throws an error if a column of that name is already present in \code{DD}.
}
