% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_atable.R
\docType{package}
\name{atable-package}
\alias{atable-package}
\title{atable: Create Tables for Reporting Clinical Trials}
\description{
The packages provides functions for descriptive statistics and hypothesis tests, and arranging the results for printing.
}
\details{
The main function is \code{\link{atable}}. See documentation there.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arminstroebel/atable}
  \item Report bugs at \url{https://github.com/arminstroebel/atable/issues}
}

}
\author{
\strong{Maintainer}: Armin Ströbel \email{arminstroebel@web.de} (\href{https://orcid.org/0000-0002-6873-5332}{ORCID})

Authors:
\itemize{
  \item Alan Haynes \email{alan.haynes@ctu.unibe.ch} (\href{https://orcid.org/0000-0003-1374-081X}{ORCID})
}

}
\keyword{internal}
