% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atable_options_reset.R
\name{atable_options_reset}
\alias{atable_options_reset}
\title{Reset atable_options to default}
\usage{
atable_options_reset()
}
\description{
Does as the name implies. See also \code{\link{atable_options}}.
}
\examples{
atable_options('replace_NA_by') # show options
atable_options('replace_NA_by' = 'foo bar') # set a new value
atable_options('replace_NA_by') # show options
atable_options_reset() # restore all defaults
atable_options('replace_NA_by') # as before

}
