% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Package_atable.R
\docType{package}
\name{atable_package}
\alias{atable_package}
\title{atable: Create Tables for Reporting Clinical Trials}
\description{
The packages provides functions for descriptive statistics and hypothesis tests, and arranging the results for printing.
}
\details{
The main function is \code{\link{atable}}. See documentation there.
}
