% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_format_statistics.R
\name{check_format_statistics}
\alias{check_format_statistics}
\title{Checks the output of function format_statistics}
\usage{
check_format_statistics(x)
}
\arguments{
\item{x}{Result of function \code{format_statistics}.}
}
\value{
\code{TRUE} if \code{x} has the following properties:
\code{x} is a non-empty data.frame with 2 columns called \code{'tag'} and \code{'value'}.
Column \code{'tag'} has class factor and no duplicates.
Column \code{'value'} is a character.
Else throws an error.
}
\description{
Checks the output of function \code{\link{format_statistics}}.
}
