% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_format_tests.R
\name{check_format_tests}
\alias{check_format_tests}
\title{Checks the output of functions format_test}
\usage{
check_format_tests(x)
}
\arguments{
\item{x}{Result of function \code{format_tests}.}
}
\value{
\code{TRUE} if \code{x} has the following properties:
\code{x} is a data.frame with exactly one row and with unique colnames. Else throws an error.
}
\description{
Checks the output of function \code{\link{format_tests}}.
}
