% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_statistics.R
\name{check_statistics}
\alias{check_statistics}
\title{Checks the output of function statistics}
\usage{
check_statistics(x)
}
\arguments{
\item{x}{Result of function \code{statistics}.}
}
\value{
\code{TRUE} if \code{x} has the following properties:
\code{x} is a named list with length > 0.
The names of the list must not have duplicates.
The names may contain NA. Else an error.
}
\description{
Checks the output of function \code{\link{statistics}}.
}
