% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_tests.R
\name{check_tests}
\alias{check_tests}
\title{Checks the output of functions two_sample_htest and multi_sample_htest}
\usage{
check_tests(x)
}
\arguments{
\item{x}{Result of function \code{two_sample_htest} or \code{multi_sample_htest}.}
}
\value{
\code{TRUE} if \code{x} has the following properties:
\code{x} is a named list with length > 0.
The names of the list must not have duplicates.
The names may contain NA. Else an error.

Most hypothesis-test-functions in R like \code{\link[stats]{t.test}} or \code{\link[stats]{chisq.test}} return an object of class htest.
This object passes this checks.
Additional fields can be added to these objects and they will still pass this check.
}
\description{
Checks the output of function \code{\link{two_sample_htest}} and \code{\link{multi_sample_htest}}.
}
