% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_data.R
\docType{data}
\name{test_data}
\alias{test_data}
\title{A data.frame with random data of various classes}
\format{
A data frame with 129 rows and 11 variables:
\describe{
  \item{Split1}{A factor with 2 levels, drawn uniformly.}
  \item{Split2}{A factor with 3 levels, drawn uniformly.}
  \item{Group}{A factor with 2 levels, drawn uniformly.}
  \item{Group2}{A factor with 3 levels, drawn uniformly.}
  \item{Numeric}{A sample from the standard normal distribution.}
  \item{Numeric2}{A sample from the normal distribution with mean 4 and sd 3.}
  \item{Logical}{A Logical, drawn uniformly from \code{TRUE}, \code{FALSE} and \code{NA}.}
  \item{Factor}{A factor with 4 level drawn with weigths \code{1:1:2:2}.}
  \item{Ordered}{Class Ordered with 3 levels, drawn uniformly.}
  \item{Character}{Class character drawn uniformly from \code{c('a', 'b', '')}.}
  \item{Date}{Class Date, generated by adding \code{2001-05-25} to
  a sample of the Poisson distribution with lambda \code{42}.}

}
6 Missing values were randomly added to each of Numeric, Numeric2, Factor, Ordered, Character and Date.
}
\usage{
test_data
}
\description{
A data.frame intended for testing the atable function with random data and missing values in various classes.
}
\keyword{datasets}
